requirements.txt
click==8.1.3
numpy==1.24.2
python-dotenv==1.0.0
torch==2.1.0.dev20230324
